Create command line script to upload file to S3
fileUpload.js takes argument from command line which is the path to the
file to be uploaded.

Then we pass the file to `createReadStream` and use the resulting stream
as the `Body` in our upload params.

We also pass the file name, our bucket name, and public-read access.
We use the AWS-SDK and S3 constructor to upload the file.